requirements.txt
iniconfig==2.3.0
packaging==26.0
pillow==12.2.0
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
